Adapt say all for list items to the changes introduced in 2021.1
